Refine setAnimationInterval from double to float.

+++ cocos/platform/CCApplicationProtocol.h
@@ -101,7 +101,7 @@ public:
     * @js NA
     * @lua NA
     */
-    virtual void setAnimationInterval(double interval) = 0;
+    virtual void setAnimationInterval(float interval) = 0;
 
     /** Subclass override the function to set OpenGL context attribution instead of use default value.
     * And now can only set six attributions:redBits,greenBits,blueBits,alphaBits,depthBits,stencilBits.
